Move example gateway into the same subdirectory
as its network abstraction
