+++ examples/BUILD.bazel
@@ -8,7 +8,6 @@ go_library(
     importpath = "github.com/grpc-ecosystem/grpc-gateway/examples",
     deps = [
         "//examples/gateway:go_default_library",
-        "//runtime:go_default_library",
         "@com_github_golang_glog//:go_default_library",
     ],
 )
@@ -32,6 +31,7 @@ go_test(
         "//examples/clients/abe:go_default_library",
         "//examples/clients/echo:go_default_library",
         "//examples/examplepb:go_default_library",
+        "//examples/gateway:go_default_library",
         "//examples/server:go_default_library",
         "//examples/sub:go_default_library",
         "//runtime:go_default_library",
