+++ examples/gateway/doc.go
@@ -1,0 +1,2 @@
+// Package gateway is an example of grpc-gateway server
+package gateway
