+++ examples/main.go
@@ -3,13 +3,9 @@ package main
 import (
 	"context"
 	"flag"
-	"net/http"
-	"path"
-	"strings"
 
 	"github.com/golang/glog"
 	"github.com/grpc-ecosystem/grpc-gateway/examples/gateway"
-	"github.com/grpc-ecosystem/grpc-gateway/runtime"
 )
 
 var (
@@ -18,83 +14,20 @@ var (
 	swaggerDir = flag.String("swagger_dir", "examples/examplepb", "path to the directory which contains swagger definitions")
 )
 
-func serveSwagger(w http.ResponseWriter, r *http.Request) {
-	if !strings.HasSuffix(r.URL.Path, ".swagger.json") {
-		glog.Errorf("Not Found: %s", r.URL.Path)
-		http.NotFound(w, r)
-		return
-	}
-
-	glog.Infof("Serving %s", r.URL.Path)
-	p := strings.TrimPrefix(r.URL.Path, "/swagger/")
-	p = path.Join(*swaggerDir, p)
-	http.ServeFile(w, r, p)
-}
-
-// allowCORS allows Cross Origin Resoruce Sharing from any origin.
-// Don't do this without consideration in production systems.
-func allowCORS(h http.Handler) http.Handler {
-	return http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
-		if origin := r.Header.Get("Origin"); origin != "" {
-			w.Header().Set("Access-Control-Allow-Origin", origin)
-			if r.Method == "OPTIONS" && r.Header.Get("Access-Control-Request-Method") != "" {
-				preflightHandler(w, r)
-				return
-			}
-		}
-		h.ServeHTTP(w, r)
-	})
-}
-
-func preflightHandler(w http.ResponseWriter, r *http.Request) {
-	headers := []string{"Content-Type", "Accept"}
-	w.Header().Set("Access-Control-Allow-Headers", strings.Join(headers, ","))
-	methods := []string{"GET", "HEAD", "POST", "PUT", "DELETE"}
-	w.Header().Set("Access-Control-Allow-Methods", strings.Join(methods, ","))
-	glog.Infof("preflight request for %s", r.URL.Path)
-	return
-}
-
-// Run starts a HTTP server and blocks forever if successful.
-func Run(ctx context.Context, address string, opts ...runtime.ServeMuxOption) error {
-	ctx, cancel := context.WithCancel(ctx)
-	defer cancel()
-
-	mux := http.NewServeMux()
-	mux.HandleFunc("/swagger/", serveSwagger)
-
-	gw, err := gateway.NewGateway(ctx, *network, *endpoint, opts)
-	if err != nil {
-		return err
-	}
-	mux.Handle("/", gw)
-
-	s := &http.Server{
-		Addr:    address,
-		Handler: allowCORS(mux),
-	}
-	go func() {
-		<-ctx.Done()
-		glog.Infof("Shutting down the http server")
-		if err := s.Shutdown(context.Background()); err != nil {
-			glog.Errorf("Failed to shutdown http server: %v", err)
-		}
-	}()
-
-	glog.Infof("Starting listening at %s", address)
-	if err := s.ListenAndServe(); err != http.ErrServerClosed {
-		glog.Errorf("Failed to listen and serve: %v", err)
-		return err
-	}
-	return nil
-}
-
 func main() {
 	flag.Parse()
 	defer glog.Flush()
 
 	ctx := context.Background()
-	if err := Run(ctx, ":8080"); err != nil {
+	opts := gateway.Options{
+		Addr: ":8080",
+		GRPCServer: gateway.Endpoint{
+			Network: *network,
+			Addr:    *endpoint,
+		},
+		SwaggerDir: *swaggerDir,
+	}
+	if err := gateway.Run(ctx, opts); err != nil {
 		glog.Fatal(err)
 	}
 }
