+++ examples/main_test.go
@@ -9,8 +9,22 @@ import (
 	"time"
 
 	"github.com/golang/glog"
+	"github.com/grpc-ecosystem/grpc-gateway/examples/gateway"
 	server "github.com/grpc-ecosystem/grpc-gateway/examples/server"
+	gwruntime "github.com/grpc-ecosystem/grpc-gateway/runtime"
 )
+
+func runGateway(ctx context.Context, addr string, opts ...gwruntime.ServeMuxOption) error {
+	return gateway.Run(ctx, gateway.Options{
+		Addr: addr,
+		GRPCServer: gateway.Endpoint{
+			Network: *network,
+			Addr:    *endpoint,
+		},
+		SwaggerDir: *swaggerDir,
+		Mux:        opts,
+	})
+}
 
 func runServers(ctx context.Context) <-chan error {
 	ch := make(chan error, 2)
@@ -20,7 +34,7 @@ func runServers(ctx context.Context) <-chan error {
 		}
 	}()
 	go func() {
-		if err := Run(ctx, ":8080"); err != nil {
+		if err := runGateway(ctx, ":8080"); err != nil {
 			ch <- fmt.Errorf("cannot run gateway service: %v", err)
 		}
 	}()
